AuthenticationConverter 및 AuthenticationInterceptor 추상화
