+++ src/main/java/nextstep/auth/authentication/interceptor/AuthenticationConverter.java
@@ -1,0 +1,9 @@
+package nextstep.auth.authentication.interceptor;
+
+import nextstep.auth.authentication.AuthenticationToken;
+
+import javax.servlet.http.HttpServletRequest;
+
+public interface AuthenticationConverter {
+    AuthenticationToken convert(HttpServletRequest request);
+}
