+++ src/main/java/nextstep/auth/authentication/interceptor/AuthenticationInterceptor.java
@@ -1,0 +1,53 @@
+package nextstep.auth.authentication.interceptor;
+
+import nextstep.auth.authentication.AuthenticationException;
+import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.context.Authentication;
+import nextstep.domain.member.domain.LoginMember;
+import nextstep.domain.member.service.UserDetailsService;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
+
+public abstract class AuthenticationInterceptor implements HandlerInterceptor, AuthenticationConverter{
+    private UserDetailsService userDetailsService;
+
+
+    public AuthenticationInterceptor() {
+    }
+
+    public AuthenticationInterceptor(UserDetailsService userDetailsService) {
+        this.userDetailsService = userDetailsService;
+    }
+
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
+        AuthenticationToken token = convert(request);
+        Authentication authentication = authenticate(token);
+        afterAuthentication(request, response, authentication);
+        return false;
+    }
+
+
+    public Authentication authenticate(AuthenticationToken token) {
+        LoginMember loginMember = userDetailsService.loadUserByUsername(token.getPrincipal());
+        checkAuthentication(loginMember, token);
+        return new Authentication(loginMember);
+    }
+
+    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+        if (userDetails == null) {
+            throw new AuthenticationException();
+        }
+        if (!userDetails.checkPassword(token.getCredentials())) {
+            throw new AuthenticationException();
+        }
+    }
+
+    public abstract void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException;
+}
